backend/requirements.txt
# ── Web Framework ──────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ── Database ───────────────────────────────
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0          # async PostgreSQL driver
psycopg2-binary>=2.9.9   # sync driver (Alembic 전용)

# ── Validation / Settings ──────────────────
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0

# ── Auth ───────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ── Redis ──────────────────────────────────
redis[hiredis]>=5.0.0

# ── Google Sheets (Wave 1 import) ──────────
google-auth>=2.28.0
google-api-python-client>=2.120.0
gspread>=6.0.0

# ── Excel (Wave 1 import) ──────────────────
openpyxl>=3.1.0

# ── Utilities ──────────────────────────────
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.1

# ── Testing ────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0            # TestClient
